move [l]gamma, [l]beta and lfact to SpecialFunctions.jl fix #27459
