+++ base/exports.jl
@@ -260,7 +260,6 @@ export
     floor,
     fma,
     frexp,
-    gamma,
     gcd,
     gcdx,
     hypot,
@@ -284,8 +283,6 @@ export
     ldexp,
     leading_ones,
     leading_zeros,
-    lfact,
-    lgamma,
     log,
     log10,
     log1p,
@@ -347,10 +344,6 @@ export
     ∛,
     ≈,
     ≉,
-
-# specfun
-    beta,
-    lbeta,
 
 # arrays
     axes,
